Backend role based authorization

See merge request hacktyki-2020/zespol1-laczone-samochody!38

+++ backend/src/main/java/com/euvic/carrental/controllers/RentController.java
@@ -2,6 +2,7 @@ package com.euvic.carrental.controllers;
 
 import com.euvic.carrental.model.*;
 import com.euvic.carrental.responses.DateFromDateTo;
+import com.euvic.carrental.responses.EndRentDTO;
 import com.euvic.carrental.responses.RentDTO;
 import com.euvic.carrental.responses.RentPermitRejectDTO;
 import com.euvic.carrental.services.*;
@@ -33,18 +34,51 @@ public class RentController {
     }
 
     //ADMIN
+    @RequestMapping(method = RequestMethod.GET, value = "/a/rent/active_rents")
+    public ResponseEntity<?> getActiveRents() {
+        return ResponseEntity.ok(rentService.getAllActiveRents());
+    }
 
     @RequestMapping(method = RequestMethod.GET, value = "/a/rent/pending")
     public ResponseEntity<?> getPendingRents() {
         return ResponseEntity.ok(rentService.getAllPendingRents());
     }
 
+    @RequestMapping(method = RequestMethod.GET, value = "/a/rent/car_history/{licensePlate}")
+    public ResponseEntity<?> checkCarHistory(@PathVariable final String licensePlate) {
+        return ResponseEntity.ok(rentHistoryService.getAllDTOsByCar(carService.getOnCompanyEntityByLicensePlate(licensePlate)));
+    }
+
+    @RequestMapping(method = RequestMethod.PUT, value = "/a/rent/change_car_in_rent/{id}")
+    public ResponseEntity<?> editRent(@PathVariable final Long id, @RequestBody final String licensePlate) {
+
+        int responseCode;
+        String message;
+        try {
+            final Rent rent = rentService.getEntityById(id);
+            final Car car = carService.getOnCompanyEntityByLicensePlate(licensePlate);
+            if (car == null)
+                throw new NullPointerException();
+            //TODO dodaj walidację czy samochód jest dostępny w terminie tego renta
+            rent.setCar(car);
+            rentService.addEntityToDB(rent);
+            responseCode = 200;
+            message = "Ok";
+        } catch (final NullPointerException e) {
+            responseCode = 400;
+            message = "Rent or car doesn't exist";
+        }
+
+        return ResponseEntity.status(responseCode).body(message);
+    }
+
+
     @RequestMapping(method = RequestMethod.PUT, value = "/a/rent/permit/{id}")
     public ResponseEntity<?> permitRent(@PathVariable final Long id, @RequestBody final RentPermitRejectDTO rentPermitRejectDTO) {
         final Rent rent = rentService.getEntityById(id);
         int responseCode;
         String message;
-        final Car car = carService.getEntityByLicensePlate(rentPermitRejectDTO.getLicensePlate());
+        final Car car = carService.getOnCompanyEntityByLicensePlate(rentPermitRejectDTO.getLicensePlate());
         if (car != null) {
             try {
                 rent.setIsActive(true);
@@ -69,18 +103,25 @@ public class RentController {
         int responseCode;
         String message;
         try {
-
-            final ParkingHistory parkingFrom = new ParkingHistory(null, rent.getParkingFrom());
-            final ParkingHistory parkingTo = new ParkingHistory(null, rent.getParkingTo());
-            final RentHistory rentHistory = new RentHistory(null, rent.getUser(), rent.getCar(), rent.getDateFrom(), rent.getDateTo(), parkingFrom
-                    , parkingTo, false, false, rent.getComment(), rentPermitRejectDTO.getResponse());
-            parkingHistoryService.addEntityToDB(parkingFrom);
-            parkingHistoryService.addEntityToDB(parkingTo);
-            rentHistoryService.addEntityToDB(rentHistory);
-            rentService.deleteRent(rent);
-
-            responseCode = 200;
-            message = "ok";
+            if (!rent.getIsActive()) {
+                final ParkingHistory parkingFrom = new ParkingHistory(null, rent.getParkingFrom());
+                final ParkingHistory parkingTo = new ParkingHistory(null, rent.getParkingTo());
+                final RentHistory rentHistory = new RentHistory(null, rent.getUser(), rent.getCar(), rent.getDateFrom(), rent.getDateTo(), parkingFrom
+                        , parkingTo, true, false, rent.getComment(), rentPermitRejectDTO.getResponse(), "");
+                parkingHistoryService.addEntityToDB(parkingFrom);
+                parkingHistoryService.addEntityToDB(parkingTo);
+                rentHistoryService.addEntityToDB(rentHistory);
+                final Long parkingFromId = rent.getParkingFrom().getId();
+                final Long parkingToId = rent.getParkingTo().getId();
+                rentService.deleteRent(rent);
+                parkingService.deleteParkingById(parkingFromId);
+                parkingService.deleteParkingById(parkingToId);
+                responseCode = 200;
+                message = "ok";
+            } else {
+                responseCode = 400;
+                message = "Rent request doesn't exist";
+            }
         } catch (final NullPointerException e) {
             message = "Invalid rent ID";
             responseCode = 400;
@@ -112,7 +153,7 @@ public class RentController {
     @RequestMapping(method = RequestMethod.POST, value = "/e/rent/{licensePlate}")
     public ResponseEntity<?> addRent(@PathVariable final String licensePlate, @RequestBody final RentDTO rentDTO) {
         final User user = userService.getEntityByLogin(SecurityContextHolder.getContext().getAuthentication().getName());
-        final Car car = carService.getEntityByLicensePlate(licensePlate);
+        final Car car = carService.getOnCompanyEntityByLicensePlate(licensePlate);
 
         int responseCode;
         final Long id;
@@ -120,8 +161,7 @@ public class RentController {
         try {
             if (rentService.checkMyRentsBeforeAddNewRent(rentDTO)) {
                 id = parkingService.addEntityToDB(parkingService.mapRestModel(null, rentDTO.getParkingDTOTo()));
-                car.setParking(parkingService.getEntityById(id));
-                final Rent rent = new Rent(null, user, car, rentDTO.getDateFrom(), rentDTO.getDateTo(), car.getParking(), parkingService.getEntityById(id), false, rentDTO.getComment(), "");
+                final Rent rent = new Rent(null, user, car, rentDTO.getDateFrom(), rentDTO.getDateTo(), car.getParking(), parkingService.getEntityById(id), false, rentDTO.getComment(), "", "");
                 rentService.addEntityToDB(rent);
                 responseCode = 200;
                 message = "Ok";
@@ -147,7 +187,11 @@ public class RentController {
 
         try {
             if (rent.getUser().equals(user)) {
+                final Long parkingFromId = rent.getParkingFrom().getId();
+                final Long parkingToId = rent.getParkingTo().getId();
                 rentService.deleteRent(rent);
+                parkingService.deleteParkingById(parkingFromId);
+                parkingService.deleteParkingById(parkingToId);
                 responseCode = 200;
                 message = "Ok";
             } else {
@@ -159,6 +203,61 @@ public class RentController {
             responseCode = 400;
             message = "Rent doesn't exist";
         }
+        return ResponseEntity.status(responseCode).body(message);
+    }
+
+    @RequestMapping(method = RequestMethod.DELETE, value = "/e/rent/end_rent/{id}")
+    public ResponseEntity<?> endRent(@PathVariable final Long id, @RequestBody final EndRentDTO endRentDTO) {
+        final Rent rent = rentService.getEntityById(id);
+        final User user = userService.getEntityByLogin(SecurityContextHolder.getContext().getAuthentication().getName());
+        int responseCode;
+        String message;
+
+        if (rent.getIsActive()) {
+            try {
+                if (rent.getUser().equals(user)) {
+                    final ParkingHistory parkingFrom = new ParkingHistory(null, rent.getParkingFrom());
+                    final ParkingHistory parkingTo;
+
+                    if (endRentDTO.getParkingDTO() == null) {
+                        parkingTo = new ParkingHistory(null, rent.getParkingTo());
+                        rent.getCar().setParking(rent.getParkingTo());
+                    } else {
+                        parkingTo = new ParkingHistory(null, endRentDTO.getParkingDTO());
+                        final Parking carParking = new Parking(null, endRentDTO.getParkingDTO());
+                        parkingService.addEntityToDB(carParking);
+                        rent.getCar().setParking(carParking);
+                    }
+
+                    final RentHistory rentHistory = new RentHistory(null, rent.getUser(), rent.getCar(), rent.getDateFrom(), rent.getDateTo(), parkingFrom
+                            , parkingTo, true, true, rent.getComment(), rent.getResponse(), endRentDTO.getFaultMessage());
+                    parkingHistoryService.addEntityToDB(parkingFrom);
+                    parkingHistoryService.addEntityToDB(parkingTo);
+                    rentHistoryService.addEntityToDB(rentHistory);
+                    final Long parkingFromId = rent.getParkingFrom().getId();
+                    final Long parkingToId = rent.getParkingTo().getId();
+                    rentService.deleteRent(rent);
+                    parkingService.deleteParkingById(parkingFromId);
+                    if (endRentDTO.getParkingDTO() != null)
+                        parkingService.deleteParkingById(parkingToId);
+
+                    //TODO Dodaj funkcję, która zmieni miejsce odbioru pojazdu najbliższego wypożyczenia na ten na który jest ustawiony
+                    responseCode = 200;
+                    message = "ok";
+                } else {
+                    responseCode = 403;
+                    message = "Rent is not belong to this user";
+                }
+
+            } catch (final NullPointerException e) {
+                responseCode = 401;
+                message = "Rent doesn't exist";
+            }
+        } else {
+            responseCode = 400;
+            message = "Rent doesn't exist";
+        }
+
         return ResponseEntity.status(responseCode).body(message);
     }
 }
